@@ -17,10 +17,14 @@ import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 
 export class TopicTableRowsComponent implements OnInit, OnDestroy, TableComponent {
   @Input() data: TableObject;
+  @Input() columnData: Array<any>;
+  @Input() smallTable: boolean;
 
   public topics: any;
   public paginationData: any;
   public dropdownItems = ['Edit', 'Delete'];
+  public columns: any;
+  public useSmallTable: boolean;
 
   private ngUnsubscribe: Subject<boolean> = new Subject<boolean>();
 
@@ -34,13 +38,16 @@ export class TopicTableRowsComponent implements OnInit, OnDestroy, TableComponen
   ngOnInit() {
     this.topics = this.data.data;
     this.paginationData = this.data.paginationData;
+    this.columns = this.columnData;
+    this.useSmallTable = this.smallTable;
   }
 
   deleteTopic(topic) {
     this.dialogService.addDialog(ConfirmComponent,
       {
         title: 'Delete Valued Component',
-        message: 'Click <strong>OK</strong> to delete this Topic or <strong>Cancel</strong> to return to the list.'
+        message: 'Click <strong>OK</strong> to delete this Topic or <strong>Cancel</strong> to return to the list.',
+        okOnly: true,
       }, {
         backdropColor: 'rgba(0, 0, 0, 0.5)'
       })
