@@ -7,10 +7,12 @@ import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { NgModule, APP_INITIALIZER, ApplicationRef } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { NgSelectModule } from '@ng-select/ng-select';
+import { MatTabsModule, MatMenuModule } from '@angular/material';
 
 // modules
 import { AppRoutingModule } from 'app/app-routing.module';
 import { ProjectModule } from 'app/project/project.module';
+import { ProjectNotificationModule } from 'app/project-notification/project-notification.module';
 import { ProjectsModule } from 'app/projects/projects.module';
 import { SharedModule } from './shared/shared.module';
 import { EditorModule } from '@tinymce/tinymce-angular';
@@ -18,17 +20,33 @@ import { EditorModule } from '@tinymce/tinymce-angular';
 // components
 import { AdministrationComponent } from 'app/administration/administration.component';
 import { AppComponent } from 'app/app.component';
+import { AssetTableRowsComponent } from 'app/project/compliance/submission-detail/asset-table-rows/asset-table-rows.component';
+import { AddCommentComponent } from 'app/comment-period/add-comment/add-comment.component';
+import { AddDocumentTableRowsComponent } from 'app/comment-periods/add-edit-comment-period/add-documents/add-document-table-rows/add-document-table-rows.component';
+import { AddEditCommentPeriodComponent } from 'app/comment-periods/add-edit-comment-period/add-edit-comment-period.component';
+import { AddLabelComponent } from 'app/project/project-documents/add-label/add-label.component';
 import { ConfirmComponent } from 'app/confirm/confirm.component';
+import { CommentPeriodComponent } from 'app/comment-period/comment-period.component';
+import { CommentPeriodDetailsTabComponent } from 'app/comment-period/comment-period-details-tabs/comment-period-details-tab.component';
+import { CommentPeriodsComponent } from 'app/comment-periods/comment-periods.component';
+import { CommentPeriodsTableRowsComponent } from 'app/comment-periods/comment-periods-table-rows/comment-periods-table-rows.component';
+import { AddDocumentComponent } from 'app/comment-periods/add-edit-comment-period/add-documents/add-documents.component';
 import { FooterComponent } from 'app/footer/footer.component';
+import { GetValuedComponentsComponent } from 'app/comment-period/review-comment/get-valued-components/get-valued-components.component';
+import { GetValuedComponentTableRowsComponent } from 'app/comment-period/review-comment/get-valued-components/get-valued-component-table-rows/get-valued-component-table-rows.component';
 import { HeaderComponent } from 'app/header/header.component';
 import { ToggleButtonComponent } from 'app/toggle-button/toggle-button.component';
 import { HomeComponent } from 'app/home/home.component';
 import { LoginComponent } from 'app/login/login.component';
 import { NotAuthorizedComponent } from './not-authorized/not-authorized.component';
 import { ProjectComponent } from './project/project.component';
+import { ReviewCommentComponent } from 'app/comment-period/review-comment/review-comment.component';
+import { ReviewCommentsTabComponent } from 'app/comment-period/review-comments-tab/review-comments-tab.component';
+import { ReviewCommentsTabTableRowsComponent } from 'app/comment-period/review-comments-tab/review-comments-tab-table-rows/review-comments-tab-table-rows.component';
 import { SearchComponent } from 'app/search/search.component';
 import { SidebarComponent } from 'app/sidebar/sidebar.component';
 import { TopicsComponent } from 'app/administration/topics/topics.component';
+import { UploadComponent } from 'app/project/project-documents/upload/upload.component';
 
 // services
 import { AuthenticationService } from 'app/services/authentication.service';
@@ -74,11 +92,12 @@ import { OrganizationsTableRowsComponent } from './organizations/organizations-t
 import { AddEditOrganizationComponent } from './organizations/add-edit-organization/add-edit-organization.component';
 import { LinkOrganizationComponent } from './shared/components/link-organization/link-organization.component';
 import { LinkOrganizationTableRowsComponent } from './shared/components/link-organization/link-organization-table-rows/link-organization-table-rows.component';
-import { NotificationProjectsComponent } from './notification-projects/notification-projects.component';
-import { AddEditNotificationProjectComponent } from './notification-projects/add-edit-notification-project/add-edit-notification-project.component';
+import { ProjectNotificationsComponent } from './project-notifications/project-notifications.component';
+import { AddEditProjectNotificationComponent } from './project-notifications/add-edit-project-notification/add-edit-project-notification.component';
 import { NotificationProjectService } from './services/notification-project.service';
-import { NotificationProjectTableRowsComponent } from './notification-projects/notification-projects-table-rows/notification-project-table-rows.component';
-import { NotificationProjectComponent } from './notification-project/notification-project.component';
+import { ProjectNotificationComponent } from './project-notification/project-notification.component';
+import { ProjectNotificationTableRowsComponent } from './project-notifications/project-notifications-table-rows/project-notifications-table-rows.component';
+import { ProjectNotificationDetailComponent } from './project-notification/detail/project-notification-detail.component';
 
 export function kcFactory(keycloakService: KeycloakService) {
   return () => keycloakService.init();
@@ -89,17 +108,29 @@ export function kcFactory(keycloakService: KeycloakService) {
     ActivityComponent,
     ActivityDetailTableRowsComponent,
     ActivityTableRowsComponent,
+    AddCommentComponent,
+    AddDocumentComponent,
+    AddDocumentTableRowsComponent,
     AddEditActivityComponent,
+    AddEditCommentPeriodComponent,
     AddEditContactComponent,
-    AddEditNotificationProjectComponent,
+    AddEditProjectNotificationComponent,
     AddEditOrganizationComponent,
     AddEditTopicComponent,
+    AddLabelComponent,
     AdministrationComponent,
     AppComponent,
+    AssetTableRowsComponent,
+    CommentPeriodComponent,
+    CommentPeriodDetailsTabComponent,
+    CommentPeriodsComponent,
+    CommentPeriodsTableRowsComponent,
     ConfirmComponent,
     ContactsComponent,
     DayCalculatorModalComponent,
     FooterComponent,
+    GetValuedComponentsComponent,
+    GetValuedComponentTableRowsComponent,
     GroupTableRowsComponent,
     HeaderComponent,
     HomeComponent,
@@ -108,13 +139,17 @@ export function kcFactory(keycloakService: KeycloakService) {
     MapComponent,
     MetricsComponent,
     NotAuthorizedComponent,
-    NotificationProjectComponent,
-    NotificationProjectsComponent,
-    NotificationProjectTableRowsComponent,
+    ProjectNotificationDetailComponent,
+    ProjectNotificationsComponent,
+    ProjectNotificationTableRowsComponent,
+    ProjectNotificationComponent,
     OrganizationsComponent,
     OrganizationsTableRowsComponent,
     PinsTableRowsComponent,
     ProjectComponent,
+    ReviewCommentComponent,
+    ReviewCommentsTabComponent,
+    ReviewCommentsTabTableRowsComponent,
     CACTableRowsComponent,
     SearchComponent,
     SearchHelpComponent,
@@ -122,6 +157,7 @@ export function kcFactory(keycloakService: KeycloakService) {
     ToggleButtonComponent,
     TopicsComponent,
     TopicTableRowsComponent,
+    UploadComponent,
     UserTableRowsComponent
   ],
   imports: [
@@ -131,13 +167,17 @@ export function kcFactory(keycloakService: KeycloakService) {
     ReactiveFormsModule,
     HttpClientModule,
     ProjectModule,
+    ProjectNotificationModule,
     ProjectsModule,
+    ProjectNotificationModule,
     SharedModule,
     EditorModule,
     AppRoutingModule, // <-- module import order matters - https://angular.io/guide/router#module-import-order-matters
-    NgbModule.forRoot(),
+    NgbModule,
     BootstrapModalModule.forRoot({ container: document.body }),
-    NgSelectModule
+    NgSelectModule,
+    MatMenuModule,
+    MatTabsModule,
   ],
   providers: [
     KeycloakService,
@@ -174,23 +214,31 @@ export function kcFactory(keycloakService: KeycloakService) {
   entryComponents: [
     ActivityDetailTableRowsComponent,
     ActivityTableRowsComponent,
+    AddDocumentTableRowsComponent,
     AddEditActivityComponent,
     AddEditContactComponent,
     AddEditOrganizationComponent,
     AddEditTopicComponent,
+    AssetTableRowsComponent,
+    CommentPeriodComponent,
+    CommentPeriodsComponent,
+    CommentPeriodsTableRowsComponent,
     ConfirmComponent,
     DayCalculatorModalComponent,
+    GetValuedComponentTableRowsComponent,
     GroupsTableRowsComponent,
     ApplicationSortTableRowsComponent,
     GroupTableRowsComponent,
     InputModalComponent,
     LinkOrganizationComponent,
     LinkOrganizationTableRowsComponent,
-    NotificationProjectTableRowsComponent,
+    ProjectNotificationTableRowsComponent,
     OrganizationsTableRowsComponent,
     PinsTableRowsComponent,
+    ReviewCommentsTabTableRowsComponent,
     CACTableRowsComponent,
     TopicTableRowsComponent,
+    UploadComponent,
     UserTableRowsComponent
   ],
   bootstrap: [AppComponent]
