@@ -1,5 +1,6 @@
 import { Component, Input, OnInit } from '@angular/core';
 
+import { TableComponent } from 'app/shared/components/table-template/table.component';
 import { TableObject } from 'app/shared/components/table-template/table-object';
 import { Router } from '@angular/router';
 import { NavigationStackUtils } from 'app/shared/utils/navigation-stack-utils';
@@ -9,12 +10,16 @@ import { NavigationStackUtils } from 'app/shared/utils/navigation-stack-utils';
   templateUrl: './user-table-rows.component.html',
   styleUrls: ['./user-table-rows.component.scss']
 })
-export class UserTableRowsComponent implements OnInit {
+export class UserTableRowsComponent implements OnInit, TableComponent {
   @Input() data: TableObject;
+  @Input() columnData: Array<any>;
+  @Input() smallTable: boolean;
 
   public contacts: any;
   public paginationData: any;
   public dropdownItems = ['Edit', 'Delete'];
+  public columns: any;
+  public useSmallTable: boolean;
 
   constructor(
     private router: Router,
@@ -24,6 +29,8 @@ export class UserTableRowsComponent implements OnInit {
   ngOnInit() {
     this.contacts = this.data.data;
     this.paginationData = this.data.paginationData;
+    this.columns = this.columnData;
+    this.useSmallTable = this.smallTable;
   }
 
   editItem(contact) {
