@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Resolve, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
+import { Resolve, ActivatedRouteSnapshot } from '@angular/router';
 import { Observable } from 'rxjs/Observable';
 
 import { SearchService } from 'app/services/search.service';
@@ -10,8 +10,9 @@ export class InspectionDetailResolver implements Resolve<Observable<object>> {
     private searchService: SearchService
   ) { }
 
-  resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<object> {
+  resolve(route: ActivatedRouteSnapshot): Observable<object> {
     const inspectionId = route.paramMap.get('inspectionId');
-    return this.searchService.getItem(inspectionId, 'Inspection');
+    // TODO: look into enabling caching again. For now its disabled
+    return this.searchService.getItem(inspectionId, 'Inspection', false);
   }
 }
