@@ -1,5 +1,5 @@
 import { Component, OnInit, ChangeDetectorRef, OnDestroy } from '@angular/core';
-import { ActivatedRoute, Router } from '@angular/router';
+import { ActivatedRoute } from '@angular/router';
 import { Subject } from 'rxjs';
 import { Compliance } from 'app/models/compliance';
 import { Project } from 'app/models/project';
@@ -29,31 +29,36 @@ export class SubmissionDetailComponent implements OnInit, OnDestroy {
     {
       name: 'Assets',
       value: 'internalExt',
-      width: 'col-2',
+      width: '15%',
       nosort: true
     },
     {
       name: 'Caption',
       value: 'caption',
-      width: 'col-4',
+      width: '25%',
       nosort: true
     },
     {
       name: 'UTM Coordinates',
       value: 'geo',
-      width: 'col-3',
+      width: '15%',
+      nosort: true
+    },
+    {
+      name: 'Date/Time Taken',
+      value: 'date',
+      width: '20%',
       nosort: true
     },
     {
       name: 'Actions',
       value: 'actions',
-      width: 'col-3',
+      width: '25%',
       nosort: true
     }
   ];
   constructor(
     private route: ActivatedRoute,
-    private router: Router,
     public api: ApiService,
     private _changeDetectionRef: ChangeDetectorRef,
     private storageService: StorageService,
