@@ -1,4 +1,4 @@
-import { Component, Input, OnInit, ChangeDetectorRef, OnDestroy, Output, EventEmitter } from '@angular/core';
+import { Component, Input, OnInit, Output, EventEmitter } from '@angular/core';
 import { TableComponent } from 'app/shared/components/table-template/table.component';
 import { TableObject } from 'app/shared/components/table-template/table-object';
 
@@ -10,10 +10,14 @@ import { TableObject } from 'app/shared/components/table-template/table-object';
 
 export class CACTableRowsComponent implements OnInit, TableComponent {
   @Input() data: TableObject;
+  @Input() columnData: Array<any>;
+  @Input() smallTable: boolean;
   @Output() selectedCount: EventEmitter<any> = new EventEmitter();
 
   public items: any;
   public paginationData: any;
+  public columns: any;
+  public useSmallTable: boolean;
 
   constructor(
   ) { }
@@ -21,6 +25,8 @@ export class CACTableRowsComponent implements OnInit, TableComponent {
   async ngOnInit() {
     this.items = this.data.data;
     this.paginationData = this.data.paginationData;
+    this.columns = this.columnData;
+    this.useSmallTable = this.smallTable;
   }
 
   selectItem(item) {
