@@ -2,7 +2,6 @@ import { Component, Input, Output, OnInit, EventEmitter } from '@angular/core';
 
 import { TableComponent } from 'app/shared/components/table-template/table.component';
 import { TableObject } from 'app/shared/components/table-template/table-object';
-import { Router } from '@angular/router';
 
 @Component({
   selector: 'tbody[app-groups-table-rows]',
@@ -12,18 +11,23 @@ import { Router } from '@angular/router';
 
 export class GroupsTableRowsComponent implements OnInit, TableComponent {
   @Input() data: TableObject;
+  @Input() columnData: Array<any>;
+  @Input() smallTable: boolean;
   @Output() selectedCount: EventEmitter<any> = new EventEmitter();
 
   public items: any;
   public paginationData: any;
+  public columns: any;
+  public useSmallTable: boolean;
 
   constructor(
-    private router: Router
   ) { }
 
   ngOnInit() {
     this.items = this.data.data;
     this.paginationData = this.data.paginationData;
+    this.columns = this.columnData;
+    this.useSmallTable = this.smallTable;
   }
 
   selectItem(item) {
