@@ -36,33 +36,33 @@ export class AddVcComponent implements OnInit, OnDestroy {
     {
       name: '',
       value: 'check',
-      width: 'col-1',
+      width: '10%',
       nosort: true
     },
     {
       name: 'Name',
       value: 'name',
-      width: 'col-3'
+      width: '20%'
     },
     {
       name: 'Description',
       value: 'description',
-      width: 'col-5'
+      width: '40%'
     },
     {
       name: 'Type',
       value: 'type',
-      width: 'col-2'
+      width: '10%'
     },
     {
       name: 'Pillar',
       value: 'pillar',
-      width: 'col-2'
+      width: '10%'
     },
     {
       name: 'Parent',
       value: 'parent',
-      width: 'col-2',
+      width: '10%',
       nosort: true
     }
   ];
@@ -123,7 +123,7 @@ export class AddVcComponent implements OnInit, OnDestroy {
             delete vc._id;
             vc.project = this.currentProject._id;
             vc.title = vc.description;
-            itemsToAdd.push({ promise: this.valuedComponentService.addToProject(vc, this.currentProject._id).toPromise(), item: item });
+            itemsToAdd.push({ promise: this.valuedComponentService.addToProject(vc).toPromise(), item: item });
           }
         });
         this.loading = false;
