@@ -33,28 +33,28 @@ export class ValuedComponentsComponent implements OnInit, OnDestroy {
     {
       name: '',
       value: 'check',
-      width: 'col-1',
+      width: '10%',
       nosort: true
     },
     {
       name: 'Type',
       value: 'type',
-      width: 'col-3'
+      width: '20%'
     },
     {
       name: 'Title',
       value: 'title',
-      width: 'col-5'
+      width: '40%'
     },
     {
       name: 'Pillar',
       value: 'pillar',
-      width: 'col-2'
+      width: '15%'
     },
     {
       name: 'Parent',
       value: 'parent',
-      width: 'col-2',
+      width: '15%',
       nosort: true
     }
   ];
