@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, Router, Route } from '@angular/router';
+import { CanActivate, Router } from '@angular/router';
 import { Observable } from 'rxjs';
 import { KeycloakService } from 'app/services/keycloak.service';
 
@@ -11,13 +11,11 @@ export class AuthGuard implements CanActivate {
     private keycloakService: KeycloakService, ) {
   }
 
-  canActivate(next: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean> | Promise<boolean> | boolean {
+  canActivate(): Observable<boolean> | Promise<boolean> | boolean {
     if (this.keycloakService.isValidForSite()) {
       return true;
     }
     this.router.navigate(['/not-authorized']);
     return false;
   }
-
-
 }
