@@ -4,20 +4,20 @@ import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/catch';
 
 import { ApiService } from './api';
-import { NotificationProject } from 'app/models/notificationProject';
+import { ProjectNotification } from 'app/models/projectNotification';
 
 @Injectable()
 export class NotificationProjectService {
 
   constructor(private api: ApiService) { }
 
-  save(notificationProject: NotificationProject, publish: boolean = null): Observable<NotificationProject> {
+  save(notificationProject: ProjectNotification, publish: boolean = null): Observable<ProjectNotification> {
     return this.api.saveNotificationProject(notificationProject, publish)
       .catch(error => this.api.handleError(error));
   }
 
-  add(notificationProject: NotificationProject, publish: boolean = false): Observable<NotificationProject> {
-    return this.api.addNotificationProject(notificationProject, publish)
+  add(projectNotification: ProjectNotification, publish: boolean = false): Observable<ProjectNotification> {
+    return this.api.addProjectNotification(projectNotification, publish)
       .catch(error => this.api.handleError(error));
   }
 }
