@@ -1,4 +1,4 @@
-import { Component, Input, Output, OnInit, EventEmitter, OnDestroy } from '@angular/core';
+import { Component, Input, Output, OnInit, EventEmitter } from '@angular/core';
 
 import { TableComponent } from 'app/shared/components/table-template/table.component';
 import { TableObject } from 'app/shared/components/table-template/table-object';
@@ -15,11 +15,15 @@ import { Org } from 'app/models/org';
 
 export class LinkOrganizationTableRowsComponent implements OnInit, TableComponent {
   @Input() data: TableObject;
+  @Input() columnData: Array<any>;
+  @Input() smallTable: boolean;
   @Output() selectedCount: EventEmitter<any> = new EventEmitter();
 
   public organizations: any;
   public paginationData: any;
   public showCheckboxes = false;
+  public columns: any;
+  public useSmallTable: boolean;
 
   constructor(
     private router: Router,
@@ -31,6 +35,8 @@ export class LinkOrganizationTableRowsComponent implements OnInit, TableComponen
     this.organizations = this.data.data;
     this.showCheckboxes = this.storageService.state.showOrgTableCheckboxes;
     this.paginationData = this.data.paginationData;
+    this.columns = this.columnData;
+    this.useSmallTable = this.smallTable;
     this.organizations.forEach((org: Org) => {
       org.checkbox = this.storageService.state.selectedOrgs.some(element => {
         return org._id === element._id;
@@ -44,7 +50,7 @@ export class LinkOrganizationTableRowsComponent implements OnInit, TableComponen
       if (item.checkbox) {
         this.storageService.state.selectedOrgs.push(item);
       } else {
-        this.storageService.state.selectedOrgs = this.storageService.state.selectedOrgs.filter(function (value, index, arr) {
+        this.storageService.state.selectedOrgs = this.storageService.state.selectedOrgs.filter(function (value) {
           return value._id !== item._id;
         });
       }
