@@ -6,7 +6,6 @@ import { filter } from 'rxjs/operators';
 import { StorageService } from 'app/services/storage.service';
 import { Subject } from 'rxjs/Subject';
 import { KeycloakService } from 'app/services/keycloak.service';
-import { ApiService } from 'app/services/api';
 
 @Component({
   selector: 'app-sidebar',
@@ -23,7 +22,9 @@ export class SidebarComponent implements OnInit, OnDestroy {
   public showNotificationProjects = false;
   public showProjectDetails = false;
   public showProjectDetailsSubItems = false;
+  public showProjectNotificationDetails = false;
   public currentProjectId = '';
+  public mainRouteId = '';
   public currentMenu = '';
   public showArchiveButton = false;
 
@@ -36,7 +37,6 @@ export class SidebarComponent implements OnInit, OnDestroy {
     private storageService: StorageService,
     private keycloakService: KeycloakService,
     private sideBarService: SideBarService,
-    private apiService: ApiService
   ) {
 
     router.events.pipe(
@@ -65,17 +65,38 @@ export class SidebarComponent implements OnInit, OnDestroy {
     if (roles !== null && roles.includes('inspector')) {
       this.isInspectorRole = true;
     }
-
-    // Remove this when notification projects are ready.
-    if (this.apiService.env === 'demo') {
-      this.showNotificationProjects = true;
-    }
   }
 
+  /**
+   * Sets the active menu item in the sibebar.
+   */
   SetActiveSidebarItem() {
-    let urlArray = this.routerSnapshot.url.split('/');
-    // urlArray[0] will be empty so we use shift to get rid of it.
+    const urlArray = this.routerSnapshot.url.split('/');
+
+    // The first element will be empty, so shift in order to remove it.
     urlArray.shift();
+    const [mainRoute, mainRouteId, currentMenu] = urlArray;
+
+    this.mainRouteId = mainRouteId;
+    this.currentMenu = currentMenu && currentMenu.split(';')[0];
+
+    switch (mainRoute) {
+      case 'p':
+        this.showProjectDetails = true;
+        this.showProjectNotificationDetails = false;
+        break;
+
+      case 'pn':
+        this.showProjectNotificationDetails = true;
+        this.showProjectDetails = false;
+        break;
+      default:
+        // There is now sub-menu so the main route ID becomes the main route. This is a root level page.
+        this.mainRouteId = mainRoute;
+        this.showProjectNotificationDetails = false;
+        this.showProjectDetails = false;
+    }
+
     if (urlArray[0] === 'p') {
       switch (urlArray[2]) {
         // case 'compliance': {
@@ -141,6 +162,11 @@ export class SidebarComponent implements OnInit, OnDestroy {
   goToDocuments(currentProjectId) {
     this.storageService.state.projectDocumentTableParams = null;
     this.router.navigate(['p', currentProjectId, 'project-documents']);
+  }
+
+  goToPnDocuments(currentProjectId) {
+    this.storageService.state.projectDocumentTableParams = null;
+    this.router.navigate(['pn', currentProjectId, 'project-notification-documents', { notificationProjectId: currentProjectId}]);
   }
 
   ngOnDestroy() {
